+++ src/main/app/controller/PostController.ts
@@ -60,7 +60,7 @@ export class PostController<T extends AnyObject> {
   private async saveAndContinue(req: AppRequest<T>, res: Response, form: Form, formData: Partial<Case>): Promise<void> {
     // Object.assign(req.session.userCase, formData);
     req.session.errors = form.getErrors(formData);
-
+    console.log("errors are:", req.session.errors)
     this.filterErrorsForSaveAsDraft(req);
 
     if (req.session.errors.length) {
