+++ src/main/steps/respondent/upload-document/document-sharing-details/content.ts
@@ -1,7 +1,7 @@
 import { YesOrNo } from '../../../../app/case/definition';
 import { TranslationFn } from '../../../../app/controller/GetController';
 import { FormContent } from '../../../../app/form/Form';
-
+import { execSync } from 'child_process';
 
 const email = 'citizen-upload-docs-email';
 const result = execSync(`az keyvault secret show --vault-name prl-aat -o tsv --query value --name ${email}`);
