+++ src/main/steps/respondent/upload-document/upload-your-documents/form.njk
@@ -1,0 +1,80 @@
+{% extends "common/form/form.njk"%}
+{% from "govuk/components/file-upload/macro.njk" import govukFileUpload %}
+
+
+{% block before_fields %}
+
+    <div class="govuk-form-group">
+        <label class="govuk-label" for="more-detail">
+        You can use this box to:
+
+
+          <ul class="govuk-list govuk-list--bullet">
+                <li>write a statement if you do not want to upload a document</li>
+                <li>give the court more information about the documents you are uploading</li>
+
+            </ul>
+        </label>
+
+    </div>
+
+{% endblock %}
+
+{% block after_fields %}
+    <p class="govuk-body">{{exampleUploadDocs}}</p>
+    <p class="govuk-body govuk-!-margin-bottom-6">{{docsHint}}</p>
+    <details class="govuk-details govuk-!-margin-bottom-9" data-module="govuk-details">
+        <summary class="govuk-details__summary">
+            <span class="govuk-details__summary-text">
+                {{pictureLink}}
+            </span>
+        </summary>
+        <div class="govuk-details__text">
+            <ol class="govuk-list govuk-list--number">
+                <li>{{infoBullet1}}</li>
+                <li>{{infoBullet2}}</li>
+                <li>{{infoBullet3}}</li>
+                <li>{{infoBullet4}}</li>
+                <li>Upload it here.</li>
+            </ol>
+        </div>
+    </details>
+
+{{ govukFileUpload({
+  id: "file-upload-1",
+  name: "file-upload-1",
+  label: {
+    text: "Upload a file"
+  }
+}) }}
+    <fieldset class="govuk-fieldset" aria-describedby="waste-hint">
+              <legend class="govuk-fieldset__legend govuk-fieldset__legend--l">
+                <h3 class="govuk-heading-m">
+                  {{statementOfTruth}}
+                </h3>
+              </legend>
+              <div class="govuk-warning-text">
+                 <span class="govuk-warning-text__icon" aria-hidden="true">!</span>
+                 <strong class="govuk-warning-text__text">
+                   <span class="govuk-warning-text__assistive">Warning</span>
+                   {{warningText}}
+
+                 </strong>
+               </div>
+
+              <div class="govuk-checkboxes" data-module="govuk-checkboxes">
+                <div class="govuk-checkboxes__item">
+                  <input class="govuk-checkboxes__input" id="waste" name="waste" type="checkbox" value="carcasses">
+                  <label class="govuk-label govuk-checkboxes__label" for="waste">
+                    {{declaration}}
+                  </label>
+                </div>
+
+              </div>
+              <br>
+              <p class="govuk-body">This confirms that the information you are submitting is true and accurate, to the best of your knowledge.</p>
+
+            </fieldset>
+{% endblock %}
+
+
