+++ src/main/steps/respondent/upload-document/upload-your-documents/template.njk
@@ -1,0 +1,140 @@
+
+{% extends "common/template.njk" %}
+{% from "govuk/components/file-upload/macro.njk" import govukFileUpload %}
+{% from "govuk/components/button/macro.njk" import govukButton %}
+
+<div class="govuk-form-group">
+{% block page_content %}
+<div class="govuk-form-group">
+    <h1 class="govuk-heading-xl">
+        <span class="govuk-caption-l">{{caption}}</span>
+        {{document_type}}
+    </h1>
+      {% if title %}
+    <h1 class="govuk-heading-l">{{ title }}</h1>
+  {% endif %}
+      <label class="govuk-label" for="more-detail">
+        You can use this box to:
+
+
+          <ul class="govuk-list govuk-list--bullet">
+            <li>write a statement if you do not want to upload a document</li>
+            <li>give the court more information about the documents you are uploading</li>
+
+          </ul>
+      </label>
+      <textarea class="govuk-textarea" id="more-detail" name="more-detail" rows="8"></textarea>
+      <form action="/document-manager/generatePdf?_csrf={{ csrfToken }}" method="post" novalidate="">
+
+      <button class="govuk-button upload-addnew govuk-button--secondary" data-module="govuk-button">
+        Submit
+      </button>
+
+    </form>
+
+</div>
+<p class="govuk-body">If you are uploading documents from a computer, name the files clearly. For example, letter-from-school.doc.</p>
+<p class="govuk-body govuk-!-margin-bottom-6">Files must end with JPG, BMP, PNG,TIF, PDF, DOC or DOCX.</p>
+<details class="govuk-details govuk-!-margin-bottom-9" data-module="govuk-details">
+            <summary class="govuk-details__summary">
+              <span class="govuk-details__summary-text">
+                How to take a picture of a document on your phone and upload it
+              </span>
+            </summary>
+            <div class="govuk-details__text">
+              <ol class="govuk-list govuk-list--number">
+                <li>Place your document on a flat service in a well-lit room. Use a flash if you need to.</li>
+                <li>Take a picture of the whole document. You should be able to see its edges.</li>
+                <li>Check you can read all the writing, including the handwriting.</li>
+                <li>Email or send the photo or scan to the device you are using now.</li>
+                <li>Upload it here.</li>
+              </ol>
+            </div>
+</details>
+{# <div class="uploadcomponent">
+          <div class=" govuk-!-margin-bottom-3">
+            <div class="govuk-form-group govuk-inset-text">
+              <label class="govuk-label govuk-!-font-weight-bold" for="file-upload-1">
+                Upload a file
+              </label>
+              <div class="govuk-form-group">
+                  <div class="upload hidden">
+                      <div class="govuk-form-group">
+                          <p class="govuk-body govuk-!-margin-bottom-0">
+                            <span id="upload"></span>
+                            <button class="govuk-button upload-addnew govuk-button--secondary" data-module="govuk-button">Upload file</button>
+                          </p>
+                          <div id="uploadProgressBar"></div>
+                      </div>
+                  </div>
+              </div>
+            </div>
+          </div>
+ </div> #}
+
+ 
+
+ <div id="uploadGroup" tabindex="0" class="uploadGroup">
+      <form action="/document-manager?_csrf={{ csrfToken }}" method="post" enctype="multipart/form-data">
+        {{ govukFileUpload({
+          name: "files[]",
+          attributes: {
+            multiple: true,
+            accept: "image/jpeg,image/tiff,image/png,application/pdf"
+          },
+          label: {
+            text: chooseFilePhoto
+          },
+          classes: "govuk-!-margin-bottom-4"
+        }) }}
+        {{ govukButton({
+          text: upload,
+          type: "submit",
+          classes: "govuk-button--secondary govuk-!-margin-bottom-0"
+        }) }}
+      </form>
+
+    <h2 class="govuk-heading-s">{{ uploadFiles }}</h2>
+    <p class="govuk-body govuk-section-break govuk-section-break--visible govuk-!-padding-bottom-3{% if userCase.applicantUploadFiles|length %} hidden{% endif %}" id="noFilesUploaded">{{ noFilesUploaded }}</p>
+    <ul id="filesUploaded" class="govuk-list">
+      {% for file in userCase.applicantUploadFiles %}
+        <li class="uploadedFile govuk-!-padding-top-2 govuk-!-padding-bottom-3 govuk-section-break govuk-section-break--visible">
+          {{ file.value.documentFileName }} <a href="/document-manager/delete/{{ loop.index0 }}?_csrf={{ csrfToken }}" class="govuk-link--no-visited-state">{{ remove }}</a>
+        </li>
+      {% endfor %}
+    </ul>
+  </div>
+
+ <div class="govuk-form-group">
+            <fieldset class="govuk-fieldset" aria-describedby="waste-hint">
+              <legend class="govuk-fieldset__legend govuk-fieldset__legend--l">
+                <h3 class="govuk-heading-m">
+                  Statement of truth
+                </h3>
+              </legend>
+              <div class="govuk-warning-text">
+                 <span class="govuk-warning-text__icon" aria-hidden="true">!</span>
+                 <strong class="govuk-warning-text__text">
+                   <span class="govuk-warning-text__assistive">Warning</span>
+                   Proceedings for contempt of court may be brought against anyone who makes, or causes to be made, a false statement verified by a statement of truth without an honest belief in its truth.
+
+                 </strong>
+               </div>
+              <div class="govuk-checkboxes" data-module="govuk-checkboxes">
+                <div class="govuk-checkboxes__item">
+                  <input class="govuk-checkboxes__input" id="waste" name="waste" type="checkbox" value="carcasses">
+                  <label class="govuk-label govuk-checkboxes__label" for="waste">
+                    I believe that the facts stated in these documents are true
+                  </label>
+                </div>
+
+              </div>
+              <br>
+              <p class="govuk-body">This confirms that the information you are submitting is true and accurate, to the best of your knowledge.</p>
+
+            </fieldset>
+          </div>
+
+{% endblock %}
+
+ </div>
