--- src/main/steps/respondent/upload-document/upload/template.njk
@@ -1,116 +1,0 @@
-{% extends "common/template.njk" %}
-{% from "govuk/components/file-upload/macro.njk" import govukFileUpload %}
-{% from "govuk/components/button/macro.njk" import govukButton %}
-{% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
-{% from "govuk/components/checkboxes/macro.njk" import govukCheckboxes %}
-
-{{ govukErrorSummary({
-  titleText: "There is a problem",
-  errorList: errorList
-}) }}
-
-<div class="govuk-form-group">
-  {% block page_content %}
-    <div class="govuk-form-group">
-      <h1 class="govuk-heading-xl">
-        <span class="govuk-caption-l">{{caption}}</span>
-        {{document_type}}
-      </h1>
-      {% if title %}
-        <h1 class="govuk-heading-l">{{ title }}</h1>
-      {% endif %}
-      <label class="govuk-label" for="more-detail">
-        {{line1}}
-
-        <ul class="govuk-list govuk-list--bullet">
-          <li>{{point1}}</li>
-          <li>{{point2}}</li>
-
-        </ul>
-      </label>
-      <textarea class="govuk-textarea" id="more-detail" name="more-detail" rows="8"></textarea>
-      <form action="/document-manager/generatePdf?_csrf={{ csrfToken }}" method="post" novalidate="">
-
-        <button class="govuk-button upload-addnew govuk-button--secondary" data-module="govuk-button">
-        Submit
-      </button>
-
-      </form>
-
-    </div>
-    <p class="govuk-body">{{uploadDocsLine1}}</p>
-    <p class="govuk-body govuk-!-margin-bottom-6">{{docsHint}}</p>
-    <details class="govuk-details govuk-!-margin-bottom-9" data-module="govuk-details">
-      <summary class="govuk-details__summary">
-        <span class="govuk-details__summary-text">
-          {{pictureHint}}
-        </span>
-      </summary>
-      <div class="govuk-details__text">
-        <ol class="govuk-list govuk-list--number">
-          <li>{{hint1}}</li>
-          <li>{{hint2}}</li>
-          <li>{{hint3}}</li>
-          <li>{{hint4}}</li>
-          <li>{{hint5}}</li>
-        </ol>
-      </div>
-    </details>
-    <div class="uploadcomponent">
-      <div class=" govuk-!-margin-bottom-3">
-        <div class="govuk-form-group govuk-inset-text">
-          <label class="govuk-label govuk-!-font-weight-bold" for="file-upload-1">
-                Upload a file
-              </label>
-          <div class="govuk-form-group">
-            <div class="upload hidden">
-              <div class="govuk-form-group">
-                <p class="govuk-body govuk-!-margin-bottom-0">
-                  <span id="upload"></span>
-                  <button class="govuk-button upload-addnew govuk-button--secondary" data-module="govuk-button">Upload file</button>
-                </p>
-                <div id="uploadProgressBar"></div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <div id="uploadGroup" tabindex="0" class="uploadGroup hidden">
-      <h2 class="govuk-heading-s">{{ uploadFiles }}</h2>
-      <ul id="filesUploaded" class="govuk-list">
-        {% for filename in uploadedDocsFilenames %}
-          <li class="uploadedFile govuk-!-padding-top-2 govuk-!-padding-bottom-3 govuk-section-break govuk-section-break--visible">
-            {{ filename }} aaaa <a href="/document-manager/delete/{{ loop.index0 }}?_csrf={{ csrfToken }}" class="govuk-link--no-visited-state">{{ remove }}</a>
-          </li>
-        {% endfor %}
-      </ul>
-    </div>
-
-    <div class="govuk-form-group">
-      <fieldset class="govuk-fieldset" aria-describedby="waste-hint">
-        <legend class="govuk-fieldset__legend govuk-fieldset__legend--l">
-          <h3 class="govuk-heading-m">
-                  Statement of truth
-                </h3>
-        </legend>
-        <div class="govuk-warning-text">
-          <span class="govuk-warning-text__icon" aria-hidden="true">!</span>
-          <strong class="govuk-warning-text__text">
-            <span class="govuk-warning-text__assistive">Warning</span>
-            {{warning}}
-
-          </strong>
-        </div>
-        
-        <br>
-
-      </fieldset>
-    </div>
-
-  {% endblock %}
-
-</div>
-</div>
-</div>
