+++ src/main/steps/urls.ts
@@ -51,8 +51,8 @@ export const RESPONDENT_TASK_LIST_URL: PageLink = '/respondent/task-list';
 export const UPLOAD_DOCUMENT_LIST_URL: PageLink = '/respondent/upload-document';
 export const UPLOAD_DOCUMENT_LIST_START_URL: PageLink = `${UPLOAD_DOCUMENT_LIST_URL}/start`;
 export const UPLOAD_DOCUMENT_LIST_SUMMARY_URL: PageLink = `${UPLOAD_DOCUMENT_LIST_URL}/document-sharing-details`;
-export const UPLOAD_DOCUMENT: PageLink = `${UPLOAD_DOCUMENT_LIST_URL}/upload`;
-export const UPLOAD_DOCUMENT_SUCCESS: PageLink = `${UPLOAD_DOCUMENT_LIST_URL}/success-page`;
+export const UPLOAD_DOCUMENT: PageLink = `${UPLOAD_DOCUMENT_LIST_URL}/upload-your-documents`;
+export const UPLOAD_DOCUMENT_SUCCESS: PageLink = `${UPLOAD_DOCUMENT_LIST_URL}/upload-documents-success`;
 /* Upload document end */
 /* respondent-keep-details-private */
 export const RESPONDENT: PageLink = '/respondent';
